fix(optim): cast dtypes in Cholesky lstsq fallback
The lstsq fallback raised RuntimeError("Expected input and other to have
the same dtype") whenever A and b were not strictly equal — which fired
across thousands of ICP refinement steps in BetterHumanForce/pre_optim
once Cholesky factorization failed for any reason. Cast b to A's dtype
for the solve and back to the caller's dtype for the result so the
fallback works regardless of mixed-precision input.

+++ src/better_robot/optim/solvers/cholesky.py
@@ -17,4 +17,4 @@ class Cholesky:
             L = torch.linalg.cholesky(A)
             return torch.cholesky_solve(b.unsqueeze(-1), L).squeeze(-1)
         except Exception:
-            return torch.linalg.lstsq(A, b).solution
+            return torch.linalg.lstsq(A, b.to(A.dtype)).solution.to(b.dtype)
